requirements.txt
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.11
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
numpy==1.20.1
pandas==1.2.2
progressbar2==3.53.1
pycodestyle==2.6.0
PyGithub==1.54.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-utils==2.5.6
pytz==2021.1
requests==2.25.1
six==1.15.0
smmap==3.0.5
toml==0.10.2
urllib3==1.26.3
wrapt==1.12.1
